LocalCodeBase/PublicCodeBase/NewGame_serverless_trivia/startGame/requirements.txt
azure-functions
azure-cosmos
azure-storage-blob
azure-identity